double to int

+++ main.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-int main()
+double main()
 {
 cout << "Enter A and B: ";
 
@@ -18,7 +18,7 @@ cout << "A * B = " << a * b << '\n';
 
 cout << "A / B = " << a / b << '\n';
 
-int max;
+double max;
 
         if (a > b) max = a;
